+++ tower-http/CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The implicit `async-compression` feature is removed (BREAKING) ([#642])
 - The implicit `tokio` feature is removed (BREAKING) ([#628])
 
+## Fixed
+
+- `trace`: restore failure classification at end-of-stream ([#483])
+
+[#483]: https://github.com/tower-rs/tower-http/pull/483
 [#628]: https://github.com/tower-rs/tower-http/pull/628
 [#642]: https://github.com/tower-rs/tower-http/pull/642